# internal/server/httpapi/openapi.json
{
  "components": {
    "responses": {
...
      "Conflict": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "The requested identity already exists."
      },
...
    },
    "schemas": {
      "AccountBot": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "tokens": {
            "items": {
              "$ref": "#/components/schemas/AccountToken"
            },
            "type": "array"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "created_at",
          "id",
          "tokens",
          "username"
        ],
        "type": "object"
      },
...
      "CreateBot": {
        "additionalProperties": false,
        "properties": {
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerToken": {
        "bearerFormat": "akari token",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "in": "cookie",
        "name": "akari_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "HTTP API for the akari React application, session ingest clients, and account automation. Browser requests authenticate with the akari_session cookie. Non-browser clients use a bearer token with the scope required by each operation.",
    "title": "akari API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/app/account/bots": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBot"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountBot"
                }
              }
            },
            "description": "Bot account created"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        },
        "summary": "Create a bot account",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/v1/app/account/bots/{bot_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedBotResponse"
                }
              }
            },
            "description": "Bot account soft-deleted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Soft-delete a bot account and revoke its credentials",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/v1/app/account/bots/{bot_id}/tokens": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateToken"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedTokenResponse"
                }
              }
            },
            "description": "Bot token created; the plaintext secret is returned once"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Create an API token for a bot account",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/v1/app/account/bots/{bot_id}/tokens/{token_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokedResponse"
                }
              }
            },
            "description": "Bot token revoked"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Revoke a bot account API token",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/v1/app/account/connections/{client_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokedResponse"
                }
              }
            },
            "description": "Connection revoked"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Revoke an OAuth client connection",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/v1/app/account/invites/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokedResponse"
                }
              }
            },
            "description": "Invite revoked"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Revoke a registration invitation",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/v1/app/account/overview-publication": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Publication"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationResponse"
                }
              }
            },
            "description": "Publication updated"
          }
        },
        "summary": "Publish or unpublish the current user's overview",
        "tags": [
          "Account"
        ]
      }
    },
...
    "/api/v1/app/insights": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "range",
            "schema": {
              "enum": [
                "7d",
                "30d",
                "90d",
                "year",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightsResponse"
                }
              }
            },
            "description": "Insights read model"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "503": {
            "$ref": "#/components/responses/ProjectionRebuilding"
          }
        },
        "summary": "Read quality and workflow insights",
        "tags": [
          "Application"
        ]
      }
    },
    "/api/v1/app/oauth/authorize": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "client_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "redirect_uri",
            "required": true,
            "schema": {
              "format": "uri",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "code_challenge",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resource",
            "schema": {
              "format": "uri",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthConsentResponse"
                }
              }
            },
            "description": "Consent read model"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Resolve an OAuth consent request for the signed-in user",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/v1/app/overview": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "range",
            "schema": {
              "enum": [
                "7d",
                "30d",
                "90d",
                "year",
                "all"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "user",
            "schema": {
              "items": {
                "format": "int64",
                "type": "integer"
              },
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewResponse"
                }
              }
            },
            "description": "Overview read model"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/ProjectionRebuilding"
          }
        },
        "summary": "Read fleet usage analytics",
        "tags": [
          "Application"
        ]
      }
    },
    "/api/v1/app/projects": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectsResponse"
                }
              }
            },
            "description": "Project collection"
          },
          "503": {
            "$ref": "#/components/responses/ProjectionRebuilding"
          }
        },
        "summary": "List projects with lifetime rollups",
        "tags": [
          "Application"
        ]
      }
    },
    "/api/v1/app/projects/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "range",
            "schema": {
              "enum": [
                "7d",
                "30d",
                "90d",
                "year",
                "all"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "agent",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "machine",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Project read model"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ProjectionRebuilding"
          }
        },
        "summary": "Read one project with windowed sessions and analytics",
        "tags": [
          "Application"
        ]
      }
    },
    "/api/v1/app/projects/{id}/publication": {
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Publication"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationResponse"
                }
              }
            },
            "description": "Publication updated"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Publish or unpublish a project overview",
        "tags": [
          "Application"
        ]
      }
    },
    "/api/v1/app/public/projects/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "range",
            "schema": {
              "enum": [
                "7d",
                "30d",
                "90d",
                "year",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProjectResponse"
                }
              }
            },
            "description": "Published project overview"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ProjectionRebuilding"
          }
        },
        "security": [],
        "summary": "Read a published project overview",
        "tags": [
          "Public"
        ]
      }
    },
    "/api/v1/app/public/sessions/{public_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "public_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSessionResponse"
                }
              }
            },
            "description": "Published session snapshot"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ProjectionRebuilding"
          }
        },
        "security": [],
        "summary": "Read a published session",
        "tags": [
          "Public"
        ]
      }
    },
    "/api/v1/app/public/sessions/{public_id}/transcript": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "public_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": true,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSessionResponse"
                }
              }
            },
            "description": "Published transcript page"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/ProjectionRebuilding"
          }
        },
        "security": [],
        "summary": "Read an earlier published transcript window",
        "tags": [
          "Public"
        ]
      }
    },
...
    "/api/v1/app/reparse": {
      "post": {
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReparseStatusResponse"
                }
              }
            },
            "description": "Rebuild queued"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "summary": "Queue a full projection rebuild",
        "tags": [
          "Operations"
        ]
      }
    },
...
  },
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerToken": []
    }
  ],
  "servers": [
    {
      "description": "This akari instance",
      "url": "/"
    }
  ],
...
}